+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Minor slug formatting issues.
+
 ## [3.1.2] - 2025-04-21
 
 ### Added
@@ -465,6 +471,7 @@
 
 - Public release of the Local Geohistory Project: Application repository.
 
+[Unreleased]: https://github.com/localgeohistoryproject/application/compare/v3.1.2...develop
 [3.1.2]: https://github.com/localgeohistoryproject/application/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/localgeohistoryproject/application/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/localgeohistoryproject/application/compare/v3.0.3...v3.1.0