+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Fixed
 
+- Mastodon link.
 - Slug refresh query for gis.governmentshape.
 
 ## [3.1.2] - 2025-04-21