+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
+- Map overlay layer to always use current base URL.
 - Minor slug formatting issues.
 - Split .env variable app_baseLocalGeohistoryProjectUrl into app_baseEnvironmentProjectUrl and app_baseCanonicalProjectUrl.